Routine development and re-factoring.

+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/core/core.get_account_id_by_account_code.sql
@@ -1,5 +1,5 @@
 CREATE FUNCTION core.get_account_id_by_account_code(text)
-RETURNS integer
+RETURNS bigint
 AS
 $$
 BEGIN
