+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/core/core.get_shipping_address_id_by_shipping_address_code.sql
@@ -1,5 +1,5 @@
 
-CREATE FUNCTION core.get_shipping_address_id_by_shipping_address_code(text, integer)
+CREATE FUNCTION core.get_shipping_address_id_by_shipping_address_code(text, bigint)
 RETURNS smallint
 AS
 $$
