+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/core/core.has_child_accounts.sql
@@ -1,4 +1,4 @@
-CREATE FUNCTION core.has_child_accounts(integer)
+CREATE FUNCTION core.has_child_accounts(bigint)
 RETURNS boolean
 AS
 $$
