+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/core/core.is_supplier.sql
@@ -1,4 +1,4 @@
-CREATE FUNCTION core.is_supplier(int)
+CREATE FUNCTION core.is_supplier(bigint)
 RETURNS boolean
 AS
 $$
