+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/logic/functions/transactions/transactions.get_accrued_interest-todo.sql
@@ -1,6 +1,6 @@
-DROP FUNCTION IF EXISTS transactions.get_accrued_interest(office_id integer, party_id integer);
+DROP FUNCTION IF EXISTS transactions.get_accrued_interest(office_id integer, party_id bigint);
 
-CREATE FUNCTION transactions.get_accrued_interest(office_id integer, party_id integer)
+CREATE FUNCTION transactions.get_accrued_interest(office_id integer, party_id bigint)
 RETURNS money_strict2
 AS
 $$
